test: Restructure ci checks tests
