+++ setup.cfg
@@ -6,7 +6,7 @@ max_line_length = 100
 exclude = .tox,venv
 
 [pytest]
-python_files = tests/test_*.py
+python_files = tests/test_*.py tests/**/test_*.py
 
 [isort]
 line_length = 100
